# cmd/compile: remove unnecessary MOV in channel receiving